requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.25
aiosqlite==0.19.0
alembic==1.13.1
python-dotenv==1.0.0
Pillow>=10.3.0
aiofiles==23.2.1
